Asus WL-330gE Support
Here is support for the compact Asus WL-330gE. It uses that all-in-one
5354 Broadcom chip that's also in the WL-520gU, so I assume USB
support can be soldered on. I've yet to open up the device, though.
Initial flashing is done through the CFE failsafe mode like on other
Asus devices, where you boot holding the reset button, the power LED
flashes, and you tftp an image.

Signed-off-by: David Cooper <dave@kupesoft.com>