[buildroot] make ModuleAutoLoad more readable
When doing a build with tracing on, the expansion of ModuleAutoLoad can get a
little hairy.

Using intermediate variables to name the arguments makes tracing more readable.

One side effect is that if an argument is accidentally left out, we won't get
all of the parameters shifted one over thanks to quoting (done in AutoLoad).

Signed-of-by: Philip Prindeville <philipp@redfish-solutions.com>