backfire: ar71xx: Configure WNDR3700 switch LED colors properly (backport of r23879)
For WNDR3700, use green for 1Gb/s and amber for 100Mb/s and 10Mb/s on
the four LAN port switch LEDs. The LEDs will be lit to indicate a link
and will blink to indicate activity. This matches the stock firmware
and the description printed on the label on the unit's underside.

Signed-off-by: Mark Mentovai <mark@moxienet.com>