mac80211: add WLAN LED support for Ralink RT305x SoC Devices
This patch adds WLAN LED support to the mac80211 driver for Ralink
rt2x00/rt2800 (rt305x) SoC devices. The current driver in
kmod-rt2800-lib is based upon PCI, not SoC. The WLAN LED drivers in
rt2800lib.c set the LED brightness via an MCU request, but do nothing
for SoC. This patch checks for SoC and sets the register to enable the
WLAN LED (instead of an MCU request).  This fixes the WLAN LED for
RT305x devices (such as the HW550-3G).

Signed-off-by: Layne Edwards <ledwards76@gmail.com>