ar71xx: TEW632BRP has buttons 'active low', fix board config
Looks like trunk@37090 has changed the logic that determines
if router's button was pressed. This resulted in TEW632BRP
always booting into failsafe mode because it detected pressed
button (which was not pressed).

Measure with voltmeter has shown that buttons on this router
are actually active-low. This patch reflects this fact in
board configuration and fixes 'load info failsafe mode' issue.

Note: it looks like Trendnet TEW632BRP and a close relative
to TEW-652BRP V1.0 and D-Link DIR-615C1, so same problem may
exist on those routers as well. This patch doesn't affect
routers other than 632BRP and unfortunately I do not have
hardware to test this issue with other routers.

This patch fixes #13893 and is tested on actuall Trendnet
TEW632BRP.

Signed-off-by: Nikolay Martynov <mar.kolya@gmail.com>
Signed-off-by: Gabor Juhos <juhosg@openwrt.org>