ar71xx: Add support for MikroTik RB951Ui-2HnD
I noticed that the patch at http://patchwork.openwrt.org/patch/4017/
for adding support for the MikroTik RouterBOARD 951Ui-2HnD had been
abandoned because it wasn't generated and sent to the mailing list
correctly and doesn't apply as a result.  I have cleaned up this patch.

When testing this on real hardware, I also noticed that wireless didn't
work, so this patch fixes that as well.

This patch applies cleanly to SVN 39392.

Signed-off-by: Matthew Reeve <mreeve@tenxnetworks.com>
Patchwork: http://patchwork.openwrt.org/patch/4773/
[juhosg:
  - drop the 'rb951ui_wlan_init' function and rework the code to
    use the recently introduced rb95x_wlan_init function instead,
  - fix GPIO number of the port5 LED,
  - rename LEDs according to the standard LED naming conventions,
    and use 'rb' prefix in the names]
Signed-off-by: Gabor Juhos <juhosg@openwrt.org>